+++ eqcorrscan/core/match_filter/party.py
@@ -838,7 +838,8 @@ class Party(object):
         """
         process_cores = process_cores or cores
         template_groups = group_templates(
-            [_f.template for _f in self.families])
+            [_f.template for _f in self.families
+             if len(_f) > 0])  # Fix for #341
         catalog = Catalog()
         for template_group in template_groups:
             family = [_f for _f in self.families